% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones.R
\name{cargar_generico}
\alias{cargar_generico}
\title{Carga de archivos RData}
\usage{
cargar_generico(archivo)
}
\arguments{
\item{archivo}{path al archivo RData que se quiere cargar}
}
\description{
Forma de cargar un archivo RData con un solo objeto para poder asignarlo directamente a un objeto nuevo
}
\keyword{files}
\keyword{load}
