% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones.R
\name{completar_faltantes}
\alias{completar_faltantes}
\title{Completar dummys faltantes}
\usage{
completar_faltantes(base_aevaluar, variables_onehot)
}
\arguments{
\item{base_aevaluar}{base que se quiere usar para evaluar modelo xgboost}

\item{variables_onehot}{nombre de todas las dummys que modelo xgboost espera se encuentren en la base}
}
\description{
Crea las dummys que falten en la base
}
\keyword{complete}
\keyword{dummys}
\keyword{xgboost}
