% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones.R
\name{iniciar_proyecto}
\alias{iniciar_proyecto}
\title{iniciar_proyecto}
\usage{
iniciar_proyecto(PATH = NULL, nombre = NULL, autor = NULL)
}
\arguments{
\item{PATH}{path al lugar donde se quiere iniciar un proyecto}

\item{nombre}{nombre del proyecto, carpeta a crear}

\item{autor}{nombre de quien trabajara sobre el proyecto}
}
\description{
Crea las dummys que falten en la base
}
\keyword{ais}
\keyword{inicio}
\keyword{proyecto}
