% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones.R
\name{onehot_transform}
\alias{onehot_transform}
\title{One hot encoding para variables categoricas}
\usage{
onehot_transform(base, variables_categoricas, guardar = FALSE,
  directorio = getwd(), sufijo = "")
}
\arguments{
\item{base}{data.frame a transformar sus variables}

\item{variables_categoricas}{vector de nombres de las variables categoricas de la base}

\item{guardar}{si se guardaran los nombres de las variables oneHot, defaults to FALSE}

\item{directorio}{directorio donde se guardarian los nombres, defaults to current directory}

\item{sufijo}{sufijo a anexar al nombre del archivo de nombres de variables, defaults to ""}
}
\description{
Transforma variables categoricas de una base a dummys y la retorna. Tiene la opcion de guardar los nombres de variables oneHot, necesarios para evaluar una nueva base
}
\keyword{categorical}
\keyword{transform}
