% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones.R
\name{transform2Dummy}
\alias{transform2Dummy}
\title{Transform2Dummy}
\usage{
transform2Dummy(variable, prefix = NULL)
}
\arguments{
\item{variable}{variable categorica a transformar}

\item{prefix}{prefijo a agregarle a los nombres de las dummys (formato character)}
}
\description{
Transforma una variable categorica a dummys
}
\keyword{categorical}
\keyword{transform}
