% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones.R
\name{guardar_variables_modelo}
\alias{guardar_variables_modelo}
\title{One hot encoding para variables categoricas}
\usage{
guardar_variables_modelo(base, variable_dependiente, directorio = "",
  sufijo = "")
}
\arguments{
\item{base}{data.frame a transformar sus variables}

\item{variable_dependiente}{vector de nombres de las variables categoricas de la base}

\item{directorio}{directorio donde se guardarian los nombres, defaults to current directory}

\item{sufijo}{sufijo a anexar al nombre del archivo de nombres de variables, defaults to ""}
}
\description{
Funcion re penca que nadie deberia usar para guardar las variables del modelo
}
\keyword{categorical}
\keyword{transform}
