% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones.R
\name{train_test_split}
\alias{train_test_split}
\title{tran_test_split}
\usage{
train_test_split(base, porc_val = 0.1, seed = 2017)
}
\arguments{
\item{base}{data.frame a separar en entrenamiento y validacion}

\item{porc_val}{porcentaje de la muestra que se asignará a validacion. Default 10%}

\item{seed}{semilla para que sea replicable la division}
}
\description{
Separa aleatoriamento una base en entrenamiento y validacion
}
\keyword{split}
\keyword{test}
\keyword{train}
